requirements.txt
colorama==0.4.6
pathlib==1.0.1
typing-extensions==4.7.1
python-dotenv==1.0.0
sqlalchemy==2.0.20
cryptography==41.0.2
requests==2.31.0
pyyaml==6.0.1
rich==13.5.2
python-nmap==0.7.1
impacket==0.11.0
telnetlib3==2.0.4
pytest==7.4.0
shodan==1.31.0
dnspython==2.6.1
paramiko==3.4.0
evdev==1.7.1
Pillow==10.3.0